requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
asttokens==3.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cloudinary==1.39.0
colorama==0.4.6
comm==0.2.3
cryptography==46.0.6
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
dj-database-url==3.1.2
Django==5.0.3
django-cors-headers==4.9.0
django-storages==1.14.2
djangorestframework==3.15.1
executing==2.2.1
google-ai-generativelanguage==0.4.0
google-api-core==2.25.2
google-auth==2.49.1
google-genai==1.69.0
google-generativeai==0.4.1
googleapis-common-protos==1.73.1
grpcio==1.78.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
packaging==26.0
parso==0.8.6
platformdirs==4.9.4
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==4.25.9
psutil==7.2.2
psycopg==3.1.18
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==27.1.0
requests==2.31.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
stack-data==0.6.3
tenacity==9.1.4
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
websockets==16.0
